REFACTOR: apidash_genai package renamed to genai
