+++ lib/models/generic_request_model.dart
@@ -1,4 +1,4 @@
-import 'package:apidash_genai/models/ai_request_model.dart';
+import 'package:genai/models/ai_request_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:apidash_core/apidash_core.dart';
 part 'generic_request_model.freezed.dart';
