+++ lib/providers/settings_providers.dart
@@ -1,5 +1,5 @@
 import 'package:apidash_core/apidash_core.dart';
-import 'package:apidash_genai/llm_saveobject.dart';
+import 'package:genai/llm_saveobject.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../models/models.dart';
