+++ lib/screens/common_widgets/api_type_dropdown.dart
@@ -1,11 +1,11 @@
 import 'package:apidash_core/apidash_core.dart';
 
-import 'package:apidash_genai/llm_input_payload.dart';
-import 'package:apidash_genai/llm_saveobject.dart';
-import 'package:apidash_genai/models/ai_request_model.dart';
-import 'package:apidash_genai/providers/gemini/gemini.dart';
-import 'package:apidash_genai/providers/gemini/models.dart';
-import 'package:apidash_genai/providers/providers.dart';
+import 'package:genai/llm_input_payload.dart';
+import 'package:genai/llm_saveobject.dart';
+import 'package:genai/models/ai_request_model.dart';
+import 'package:genai/providers/gemini/gemini.dart';
+import 'package:genai/providers/gemini/models.dart';
+import 'package:genai/providers/providers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:apidash/providers/providers.dart';
