+++ lib/screens/home_page/editor_pane/url_card.dart
@@ -1,11 +1,11 @@
 import 'package:apidash_core/apidash_core.dart';
 
 import 'package:apidash_design_system/apidash_design_system.dart';
-import 'package:apidash_genai/llm_input_payload.dart';
-import 'package:apidash_genai/llm_saveobject.dart';
-import 'package:apidash_genai/models/ai_request_model.dart';
+import 'package:genai/llm_input_payload.dart';
+import 'package:genai/llm_saveobject.dart';
+import 'package:genai/models/ai_request_model.dart';
 
-import 'package:apidash_genai/widgets/llm_selector.dart';
+import 'package:genai/widgets/llm_selector.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:apidash/providers/providers.dart';
