+++ lib/screens/settings_page.dart
@@ -1,4 +1,4 @@
-import 'package:apidash_genai/widgets/llm_selector.dart';
+import 'package:genai/widgets/llm_selector.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
