+++ packages/apidash_core/pubspec.yaml
@@ -23,8 +23,8 @@ dependencies:
     path: ../postman
   seed: ^0.0.3
   xml: ^6.3.0
-  apidash_genai:
-    path: ../apidash_genai
+  genai:
+    path: ../genai
 
 dev_dependencies:
   flutter_test:
