+++ packages/genai/README.md
@@ -1,4 +1,4 @@
-# apidash_genai
+# genai
 This Package contains all the code related to generative AI capabilities and is a foundational feature that other APIDash parts can make use of.
 
 ### Getting LLM Models for a given Provider
