+++ packages/genai/lib/models/ai_request_model.freezed.dart
