+++ packages/genai/lib/models/ai_response_model.dart
