+++ packages/genai/lib/models/ai_response_model.freezed.dart
