+++ packages/genai/lib/providers/azureopenai/azureopenai.dart
@@ -1,14 +1,13 @@
-import 'package:apidash_genai/llm_config.dart';
-import 'package:apidash_genai/llm_input_payload.dart';
-import 'package:apidash_genai/llm_request.dart';
-import 'package:apidash_genai/providers/common.dart';
+import 'package:genai/llm_config.dart';
+import 'package:genai/llm_input_payload.dart';
+import 'package:genai/llm_request.dart';
+import 'package:genai/providers/common.dart';
 
-class OpenAIModelController extends ModelController {
-  static final instance = OpenAIModelController();
-
+class AzureOpenAIModelController extends ModelController {
+  static final instance = AzureOpenAIModelController();
   @override
   LLMInputPayload get inputPayload => LLMInputPayload(
-    endpoint: 'https://api.openai.com/v1/chat/completions',
+    endpoint: '', //TO BE FILLED BY USER
     credential: '',
     systemPrompt: '',
     userPrompt: '',
@@ -25,12 +24,14 @@ class OpenAIModelController extends ModelController {
     LLMInputPayload inputPayload, {
     bool stream = false,
   }) {
+    if (inputPayload.endpoint.isEmpty) {
+      throw Exception('MODEL ENDPOINT IS EMPTY');
+    }
     return LLMRequestDetails(
       endpoint: inputPayload.endpoint,
-      headers: {'Authorization': "Bearer ${inputPayload.credential}"},
+      headers: {'api-key': inputPayload.credential},
       method: 'POST',
       body: {
-        'model': model.identifier,
         if (stream) ...{'stream': true},
         "messages": [
           {"role": "system", "content": inputPayload.systemPrompt},
