+++ packages/genai/pubspec.yaml
@@ -1,4 +1,4 @@
-name: apidash_genai
+name: genai
 description: "Generative AI Capabilities for APIDash"
 version: 0.0.1
 homepage:
