+++ pubspec.yaml
@@ -14,8 +14,8 @@ dependencies:
     path: packages/apidash_core
   apidash_design_system:
     path: packages/apidash_design_system
-  apidash_genai:
-    path: packages/apidash_genai
+  genai:
+    path: packages/genai
   carousel_slider: ^5.0.0
   code_builder: ^4.10.0
   csv: ^6.0.0
